Compiled Constellation to support leaving nodes
Compiled Constellation to support custom leaving nodes, the jar files
can be found in Constellation/lib/ this is not 100% reliable and needs
to be testing. The jar files should ONLY be used for this application,
as there is no fault-tolerance built in and messages will be lost if a
node leaves while there are incoming messages in transit.